Create baseball.py
This is a quick program asking a user to input how many tickets they would like to purchase for a baseball game and how much each ticket will cost plus extras for food and drinks. It will then ask if they would like to purchase more. 
this is a quick example of a while loop and a simple accumulator.